fix(api): correct id & salary mapping for MockAPI (data loaded successfully)

+++ lib/presentation/pages/employee_list_page.dart
@@ -114,7 +114,9 @@ class _EmployeeTile extends ConsumerWidget {
             ),
           );
           if (ok == true) {
-            await ref.read(employeeListProvider.notifier).deleteById(emp.id!);
+            await ref
+                .read(employeeListProvider.notifier)
+                .deleteById(emp.id! as int);
             ScaffoldMessenger.of(context)
                 .showSnackBar(const SnackBar(content: Text('Data dihapus')));
           }
